Dynamic Version of Amiparqueo
This Dynamic Version of Amiparqueo is a full webapp. It includes a login system, and all the functionalities of the original amiparqueo. It also incorporates some features that allows the users to fully customize their information and data. It also allows of the creation of a Master User account and a Dependent account with limited access.